+++ CHANGELOG.md
@@ -116,6 +116,9 @@
   The issue only affects scenarios where the same Go plugin is applied to different Route
   or Service entities.
   [#11306](https://github.com/Kong/kong/pull/11306)
+- Update the DNS client to follow configured timeouts in a more predictable manner.  Also fix a corner case in its
+  behavior that could cause it to resolve incorrectly during transient network and DNS server failures.
+  [#11386](https://github.com/Kong/kong/pull/11386)
 
 #### Admin API